+++ src/prescrypt/__init__.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
 from .compiler import py2js
+from .front.passes.types import JS, JSObject
 
-__all__ = ["py2js"]
+__all__ = ["py2js", "JS", "JSObject"]